requirements.txt
flask
flask-cors
gunicorn
python-dotenv
google-genai
PyPDF2
python-docx
